moving the williamson2 test case inital conditions to a separate file, and including this in both the explicit and rosenbrock formulations
