+++ test/Williamson2ShallowWaterExplicitTests.jl
@@ -6,41 +6,14 @@ using WriteVTK
 using Gridap
 using GridapGeosciences
 
+include("Williamson2InitialConditions.jl")
+
 # Solves the steady state Williamson2 test case for the shallow water equations on a sphere
 # of physical radius 6371220m. Involves a modified coriolis term that exactly balances
 # the potential gradient term to achieve a steady state
 # reference:
 # D. L. Williamson, J. B. Drake, J. J.HackRüdiger Jakob, P. N.Swarztrauber, (1992)
 # J Comp. Phys. 102 211-224
-
-# Constants of the Williamson2 test case
-const α  = π/4.0              # deviation of the coriolis term from zonal forcing
-const U₀ = 38.61068276698372  # velocity scale
-const H₀ = 2998.1154702758267 # mean fluid depth
-
-# Modified coriolis term
-function f₀(xyz)
-   θϕr   = xyz2θϕr(xyz)
-   θ,ϕ,r = θϕr
-   2.0*Ωₑ*( -cos(θ)*cos(ϕ)*sin(α) + sin(ϕ)*cos(α) )
-end
-
-# Initial velocity
-function u₀(xyz)
-  θϕr   = xyz2θϕr(xyz)
-  θ,ϕ,r = θϕr
-  u     = U₀*(cos(ϕ)*cos(α) + cos(θ)*sin(ϕ)*sin(α))
-  v     = -U₀*sin(θ)*sin(α)
-  spherical_to_cartesian_matrix(θϕr)⋅VectorValue(u,v,0)
-end
-
-# Initial fluid depth
-function h₀(xyz)
-  θϕr   = xyz2θϕr(xyz)
-  θ,ϕ,r = θϕr
-  h  = -cos(θ)*cos(ϕ)*sin(α) + sin(ϕ)*cos(α)
-  H₀ - (rₑ*Ωₑ*U₀ + 0.5*U₀*U₀)*h*h/g
-end
 
 l2_err_u = [0.009872849324844975, 0.002843859325502451,  0.0007415233680147055]
 l2_err_h = [0.00561048961466849,  0.0014553891676895917, 0.0003681302039168149]
